[wptpr.live] Change PR deployment to use 'pull_request_target'
The 'pull_request_target' workflow event is identical to
'pull_request', except it is dispatched in the context of the base
commit rather than the merge commit. This means that only workflows that
already exist in the repository before the PR are run, and using code
that was already in the repository - aka, the code is trustworthy!

Having trustworthy code means we can greatly simplify the design for
deploying code to wptpr.live, and also easily support PRs from forks too
- this change does both.

Fixes https://github.com/web-platform-tests/wpt.live/issues/27