[scroll-animations] Run style/layout twice if necessary
This CL solves the "first frame problem" [1], where ScrollTimelines
created by style recalc end up as inactive, because layout hasn't
run yet at the time that style recalc takes place.

When a new timeline is created, it goes into a list of "unvalidated"
timelines. During the lifecycle update, right after the style and
layout steps, we go through all such timelines and check if the current
timeline state (snapshot) is "stale" compared to a freshly computed
snapshot. If it is, it means that layout had some effect on the
timeline(s) that we would like to include in the next frame. We then
mark the associated effect targets for style recalc, which in turn
triggers a re-run of the style/layout steps in the lifecycle.

After all timelines have been validated, the list of unvalidated
timelines is cleared. This guarantees that we don't accidentally
trigger style/layout more than twice.

[1] https://github.com/w3c/csswg-drafts/issues/5261

Fixed: 1145933
Bug: 1074052
Change-Id: Ic915bb71b204c93cbde1c5f485d26a4b1c2bd09f